move connection factory to container folder
